requirements.txt
streamlit>=1.22.0,<2.0.0
langchain>=0.0.184,<0.1.0
chromadb>=0.3.21,<0.4.0
beautifulsoup4>=4.11.1,<5.0.0
requests>=2.28.1,<3.0.0
python-dotenv>=0.21.0,<1.0.0
PyPDF2>=2.11.1,<3.0.0
reportlab>=3.6.12,<4.0.0
pyyaml>=6.0,<7.0.0
groq>=0.3.0,<0.4.0